[HUDI-1222] Introduce MergeHelper.UpdateHandler as independent class
