+++ hudi-client/src/main/java/org/apache/hudi/table/action/commit/UpdateHandler.java
@@ -1,0 +1,48 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.hudi.table.action.commit;
+
+import org.apache.avro.generic.GenericRecord;
+import org.apache.hudi.common.util.queue.BoundedInMemoryQueueConsumer;
+import org.apache.hudi.io.HoodieMergeHandle;
+
+/**
+ * Consumer that dequeues records from queue and sends to Merge Handle.
+ */
+public class UpdateHandler  extends BoundedInMemoryQueueConsumer<GenericRecord, Void> {
+
+  private final HoodieMergeHandle upsertHandle;
+
+  public UpdateHandler(HoodieMergeHandle upsertHandle) {
+    this.upsertHandle = upsertHandle;
+  }
+
+  @Override
+  protected void consumeOneRecord(GenericRecord record) {
+    upsertHandle.write(record);
+  }
+
+  @Override
+  protected void finish() {}
+
+  @Override
+  protected Void getResult() {
+    return null;
+  }
+}
